refactor: decompose workflow runtime services
